Moved Integer+Bytes into its own file within utils
